Added header and event card layout, add user info
Updated the planner header to include user info and a logout button, with improved styling and responsive design. Refactored event list to use a grid layout so the whole page won't change even if there are more than 3 events, adjusted event card sizing for better look and added hover effects for cards. Also added a new event to the events data to check if the grid by 3 works.